requirements.txt
aphyt==0.1.6
cython==0.29.26
libscrc==1.7
matplotlib==3.1.2
numpy==1.21.4
pandas==1.3.5
pyserial==3.5
PySimpleGUI==4.55.1
seaborn==0.11.2
techmanpy==1.2
